+++ src/Manage_Professors/View_Professors.java
@@ -70,14 +70,17 @@ public class View_Professors {
                     String id = tblModel.getValueAt(i,0).toString();
 
                     try {
-                        pst = db.connection.prepareStatement("DELETE FROM users WHERE id_user =?");
-                        pst.setString(1,id);
-                        pst.executeUpdate();
+                        int result = JOptionPane.showConfirmDialog((Component) null, "Are you sure?",
+                                "alert", JOptionPane.OK_CANCEL_OPTION);
+                        if (result == JOptionPane.OK_OPTION) {
+                            pst = db.connection.prepareStatement("DELETE FROM users WHERE id_user =?");
+                            pst.setString(1, id);
+                            pst.executeUpdate();
 
-                        showTableData();
-                        //tblModel.removeRow(level_table.getSelectedRow());
-                        JOptionPane.showMessageDialog(null,"Deleted successfully !");
-
+                            showTableData();
+                            //tblModel.removeRow(level_table.getSelectedRow());
+                            JOptionPane.showMessageDialog(null, "Deleted successfully !");
+                        }
                     } catch (SQLException | HeadlessException ex) {
                         JOptionPane.showMessageDialog(null,ex);
                     }
